# openapi/specs.yaml
openapi: 3.1.0
info:
  title: Electronic Shop API
  description: |
    A simple e-commerce API for electronic products with JWT-based authentication.
    
    ## Authentication
    This API uses JWT (JSON Web Tokens) for authentication. To access protected endpoints:
    1. First, obtain a JWT token by calling the `/api/auth/login` endpoint with valid credentials
    2. Include the token in the `Authorization` header as `Bearer <token>` for subsequent requests
    
    ## Default Users
    For testing purposes, the following users are available:
    - **Admin**: `admin@shop.com` / `admin123` (ADMIN role)
    - **User**: `user@shop.com` / `user123` (USER role)
  version: 1.0.0
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  contact:
    name: Electronic Shop API Support
    email: support@electronicshop.com
servers:
  - url: http://dev.electronicshop.com:8888
    description: Development server
  - url: https://api.electronicshop.com
    description: Production server
...
paths:
...
  /admin/products:
    get:
      tags:
        - Admin
      summary: Get all products
      description: Retrieve a list of all products in the system.
      operationId: getAllProducts
      security:
        - BearerAuth: []
      parameters: 
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/SearchText'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListDto'
        '401':
          description: Unauthorized - invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_token:
                  summary: Missing authorization token
                  value:
                    error: Access token is missing or invalid
                expired_token:
                  summary: Expired token
                  value:
                    error: Token has expired
        '403':
          description: Forbidden - valid token but insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
    post:
      operationId: createNewProduct
      tags: 
        - Admin
      security: 
        - BearerAuth: []
      summary: Create a new product
      description: |
        Create a new product in the system. Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertForm'
      responses:
        '201':
          description: Product created successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '50x':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
  /admin/products/{id}:
    put:
      operationId: updateProduct
      security: 
        - BearerAuth: []
      tags: 
        - Admin
      summary: Update an existing product
      description: |
        Update an existing product by its ID. Requires ADMIN role.
      parameters: 
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertForm'
      responses: 
        '200':
          description: Product updated successfully
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid product data"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not found"
        '50x':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
    delete:
      operationId: deleteProduct
      security: 
        - BearerAuth: []
      tags: 
        - Admin
      summary: Delete a product
      description: |
        Delete a product by its ID. Requires ADMIN role.
      parameters: 
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Product deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not found"
        '50x':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
  /admin/deals/{id}:
    put: 
      operationId: updateDealExpiration
      tags: 
        - Admin
      summary: Update the expiration date of an existing deal
      description: |
        Update the expiration date of an existing deal. Requires ADMIN role.
      security: 
        - BearerAuth: []
      parameters: 
        - $ref: '#/components/parameters/Id' 
      requestBody: 
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealUpdateExpirationForm'
      responses: 
        '200':
          description: Deal expiration updated successfully
        '400':
          description: Bad request - invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Invalid deal data"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not found"
        '50x':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
  /admin/deals/{id}/products:
    post:
      operationId: createDealProducts
      tags: 
        - Admin
      summary: Add products to a deal
      description: |
        Add products to a specific deal by its ID. Requires ADMIN role.
      security: 
        - BearerAuth: []
      parameters: 
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MultiSelectForm'
      responses:
        '201':
          description: Add products to deal successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Not found"
        '50x':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: "Internal server error"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token authentication. 
        
        Obtain a token by calling the `/api/auth/login` endpoint.
        Include the token in requests as: `Authorization: Bearer <token>`
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Invalid credentials
      description: Standard error response
...
    ProductListDto:
      allOf: 
        - $ref: '#/components/schemas/PaginationProperties'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/ProductItemDto'
    ProductItemDto:
      type: object
      required:
        - id
        - name
        - stock
        - unitPrice
      properties:
        id:
          type: string
          description: Unique identifier for the product
          example: 12345
        name:
          type: string
          description: Name of the product
          example: Smart TV 55"
        stock:
          type: integer
          description: Available stock quantity
          example: 50
        unitPrice:
          type: number
          format: double
          description: Price per unit of the product
          example: 499.99
    PaginationProperties:
      type: object
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: Current page number
          default: 1
          example: 1
        size:
          type: integer
          description: Number of items per page
          default: 10
          example: 10
        total:
          type: integer
          description: Total number of items available
          example: 100
    ProductUpsertForm:
      type: object
      required:
        - name
        - stock
        - unitPrice
      properties:
        name:
          type: string
          description: Name of the product
          example: Smart TV 55"
        stock:
          type: integer
          description: Available stock quantity
          minimum: 1
          example: 50
        unitPrice:
          type: number
          format: double
          minimum: 1
          description: Price per unit of the product
          example: 499.99
    DealUpdateExpirationForm:
      type: object
      properties:
        newStartAt:
          type: string
          format: date-time
          description: New start date and time for the deal
          example: 2024-01-15T10:30:00Z
        newEndAt:
          type: string
          format: date-time
          description: New end date and time for the deal
          example: 2024-01-31T23:59:59Z
      required: 
        - newStartAt
        - newEndAt
    MultiSelectForm:
      type: object
      properties:
        selectedItems:
          type: array
          items:
            type: string
          description: List of selected item IDs
          example: ["12345", "67890"]
      required:
        - selectedItems
  parameters:
    Page:
      name: page
      in: query
      description: Page number to retrieve
      required: false
      schema:
        type: integer
        default: 1
    Size:
      name: size
      in: query
      description: Number of items per page
      required: false
      schema:
        type: integer
        default: 10
    SearchText:
      name: searchText
      in: query
      description: Search term to filter products by name
      required: false
      schema:
        type: string
    Id:
      name: id
      in: path
      description: Unique identifier of the resource
      required: true
      schema:
        type: string